First batch of changes for the addition of WebDAV and ChangeEvent support in ILIAS.
This commit includes:

WebDAV support
---------------
- WebDAV server:
The WebDAV server is essentially working.
Missing features: The WebDAV interface is currently only working for user accounts which are authenticated by ILIAS and by LDAP. The language variables are missing.

- Mounting Webfolders.
Mounting is essentially working.
Missing features: The mount as webfolder actions are currently visible in all skins - they will become invisible in the Delos Skin. The mount as webfolder icon on the top of a repository page is currentlymissing. The help text for browsers which don't support mounting is not displayed yet. The language variables are missing.

- File objects in the repository: 
ILIAS makes now a distinction between inline files and downloadable files.
Files display now locking information in the repository and the info panel.
Hidden files are now supported.
File sizes are now reported in scales of 1024 like Windows and Mac OS X does. 
Dates are reported concisely with 'today', 'yesterday' and 'tomorrow'.
Missing features: The inline icons are missing. The language variables are missing.

- File Access settings panel:
The panel is basically working.
Missing features: The language variables and icons are missing.


ChangeEvent support
-------------------
- Objects in the repository:
For categories, files, courses and groups, tracking is working.
Tracking information is visible in the repository view and the info panels of each object.
Missing features: For other object types, tracking support needs to be added. The new tracking tables need to be merged with the existing ones. Non-Delos skins need to be able to  mark new and changed files with an icon (not just textually). The language variables are missing.

- Tracking and Learning progress settings
Change event tracking can be enabled/disabled in the settings panel. Change event tracking is enabled by default.
Missing features: language variables are missing.


git-svn-id: http://svn.ilias.de/svn/ilias/trunk@15958 21b2c9ec-7c21-0410-8b45-9bfb8ed2bfc5

+++ classes/class.ilNestedSetXML.php
@@ -1416,13 +1416,18 @@ class ilNestedSetXML
 
 
 		// FINALLY DELETE
-		$in = " IN ('";
-		$in .= implode("','", $tag_fks);
-		$in .= "')";
-		
-		$ilDB->query("DELETE FROM xmlparam WHERE tag_fk ".$in);
-		$ilDB->query("DELETE FROM xmlvalue WHERE tag_fk ".$in);
-		$ilDB->query("DELETE FROM xmltags  WHERE tag_pk ".$in);
+		// BEGIN WebDAV: Object deletion failed if no tag_fks was present.
+		if ($tag_fks != null)
+		{
+			$in = " IN ('";
+			$in .= implode("','", $tag_fks);
+			$in .= "')";
+
+			$ilDB->query("DELETE FROM xmlparam WHERE tag_fk ".$in);
+			$ilDB->query("DELETE FROM xmlvalue WHERE tag_fk ".$in);
+			$ilDB->query("DELETE FROM xmltags  WHERE tag_pk ".$in);
+		}
+		// END WebDAV Object deletion failed if no tag_fks was present.
 
 		#$ilBench->stop('NestedSet','deleteAllChildMetaData');
 		return true;
